Centralize more inline Triton kernels
